restructure + add projection matrix
